+++ Classes/Utility/Html2TextLibrary.php
@@ -1,4 +1,5 @@
 <?php
+
 /***************************************************************
 *  Copyright notice
 *
@@ -19,37 +20,31 @@
 *  This copyright notice MUST APPEAR in all copies of the script!
 ***************************************************************/
 
-require_once(PATH_tslib.'class.tslib_pibase.php');
-
-class tx_odsplaintext extends tslib_pibase {
-	var $prefixId = 'tx_odsplaintext'; // Same as class name
-	var $scriptRelPath = 'lib/class.tx_odsplaintext.php'; // Path to this script relative to the extension dir.
-	var $extKey = 'ods_plaintext'; // The extension key.
+class Html2TextLibrary extends tslib_pibase {
+	var $prefixId = 'Html2TextLibrary'; // Same as class name
+	var $scriptRelPath = 'Classes/Utility/Html2TextLibrary.php'; // Path to this script relative to the extension dir.
+	var $extKey = 'html2textlibrary'; // The extension key.
 	var $pi_checkCHash = TRUE;
-	
-	function user_cleanup($content,$conf)	{
+
+	function convert($content, $conf) {
 
 		// Include the HTML-cleaner class definition file.
-		require_once('class.html2text.inc');
+		require_once('Html2Text.php');
 
 		// Instantiate a new instance of the class. Passing the string
 		// variable automatically loads the HTML for you.
-		$h2t = new html2text($content);
+		$htmlToText = new Html2Text($content);
 
 		// The HTML is likely full of relative links, so let's specify
 		// an absolute source.
-		$h2t->set_base_url($GLOBALS['TSFE']->absRefPrefix);
+		$htmlToText->set_base_url($GLOBALS['TSFE']->absRefPrefix);
 
 		// Simply call the get_text() method for the class to convert
 		// the HTML to the plain text. Store it into the variable.
-		$text = $h2t->get_text();
-    
+		$text = $htmlToText->get_text();
+
 		return ($text);
 	}
 }
 
-
-if (defined('TYPO3_MODE') && $TYPO3_CONF_VARS[TYPO3_MODE]['XCLASS']['ext/ods_plaintext/lib/class.tx_odsplaintext.php'])	{
-	include_once($TYPO3_CONF_VARS[TYPO3_MODE]['XCLASS']['ext/ods_plaintext/lib/class.tx_odsplaintext.php']);
-}
 ?>